requirements.txt
fastapi
uvicorn
python-multipart
pymilvus
langchain-text-splitters
pdfplumber
google-genai
nest_asyncio
sqlalchemy
psycopg2-binary
pydantic
langchain-core
langgraph
langchain-litellm
litellm
PyJWT
python-dotenv
langchain-openai
langchain-milvus
langchain-experimental
langchain-community
pypdf
python-docx
unstructured[pdf]
pillow
python-pptx
docx2txt